+++ service/controller/resource/clusterconfigmap/desired.go
@@ -68,6 +68,24 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) ([]*cor
 		}
 	}
 
+	var provider string
+	{
+		infrastructureRef := cr.Spec.InfrastructureRef
+
+		switch infrastructureRef.Kind {
+		case "AWSCluster":
+			provider = "aws"
+			break
+
+		case "AzureCluster":
+			provider = "azure"
+			break
+
+		default:
+			provider = "unknown"
+		}
+	}
+
 	configMapSpecs := []configMapSpec{
 		{
 			Name:      key.ClusterConfigMapName(&cr),
@@ -95,6 +113,7 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) ([]*cor
 				"clusterDNSIP": r.dnsIP,
 				"clusterID":    key.ClusterID(&cr),
 				"clusterCIDR":  clusterCIDR,
+				"provider":     provider,
 			},
 		},
 	}