+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add provider to the cluster values configmap.
+
 ## [0.5.0] - 2021-08-30
 
 ### Added